requirements.txt
# Web Framework and ASGI server
fastapi>=0.115.8
uvicorn>=0.34.0

# Database
sqlalchemy[asyncio]>=2.0.38
asyncpg>=0.29.0

# Data Validation
pydantic>=2.10.6
pydantic-settings>=2.1.0

# Authentication and Security
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.20

# Caching
fastapi-cache2[redis]>=0.2.2
redis<5.0.0
